src/motor/requirements.txt
RPi.GPIO==0.7.1
pyyaml==6.0.1